backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==4.1.2
djangorestframework==3.14.0
django-extensions==3.2.1
django-cors-headers==3.13.0
# Editable install with no version control (progressbar==2.5)
progressbar==2.5